requirement.txt
google-generativeai==0.8.3
python-dotenv==1.0.1
Flask==3.1.0